Added iteration 1 of the control box
This control box enable to power the Raspberry Pi from the PSU of the
Ender 3 and control the power of the printer itself, the power of a LED
stripe and fans for the control box (cooling the RPi and the Buck
converters).

The Raspberry Pi control those using a Google AIY Voice Kit V1 (I use
the servo control) but it can be done directly from the GPIO of the
voice kit (just needs more wiring logic). I also intend to add an
auxiliary box for the mic and the speaker of the Voice Kit so I can
control my printer by voice :)

Fixing is done using a mix of snap fit feet and brass inserts.