Initial redesign of the control box.
This design is still pretty rough but should be printer for test. The
front part is probably the first one to print since it can change the
design of everything else. Second part to print would be the upper
layer.

The first thing to verify is that this does not prevent the Z axis from
going all the way down.