Add some interface layers to the support
Theses interface layers draw a non continous part to ensure they do not
stick hard on the part itself.

Also increase the gap between support and wall as 0.5 is too close to
the extrusion width and was fusing with the wall in some case.

This make the support easy to remove.