Add the possibility of putting a BTT UPS Mini 24V
This replace one of the relay switch. In my case this
is replacing the relay switch for the printer as
I ended up always turning off the raspberry pi with the
printer.

Note: not yet printed, needs to be printed for correct testing. I intend
to print it together with the reverse version.