Adapt the wire guide to guide upward
Guiding on the back was a dummy move as it would collide with the frame
when on the left or right. This change also remove the DC-DC converter
case, mine just go fine with electric tape between the X-carriage and
the BMG extruder. I might consider doing an independent case that
just slide directly there later.